Simplify hash logic and update ffmpeg footer version
Simplified the hash calculation in `IsConvertedByMetadata` by removing the need to calculate and check two different hashes. Now, only one hash using the file's name and length is calculated and checked against `_convertedHashes`.

Updated the `IsConverted` method to check for a newer version of the ffmpeg footer, changing from "Lavf58.45.100" to "Lavf60.16.100".