Add support for 3gp and vob video media types
The changes introduce support for two new video media types in the
`MediaConverter.Core` namespace: `ThirdGenerationPartnershipProject`
(3gp) and `Mpeg2ProgramStream` (vob).

In `ConverterCore.cs`, the new media types are mapped to their
respective video codecs: `ThirdGenerationPartnershipProject` is mapped
to `h264`, and `Mpeg2ProgramStream` is mapped to `mpeg2video`.

In `MediaTypes.cs`, constants for the new media types are defined:
`ThirdGenerationPartnershipProject` is set to "3gp" and `Mpeg2ProgramStream`
is set to "vob". These new media types are also included in the
`AsEnumerable` method, which returns a collection of all supported
media types.